AI-PET-Agent/backend/requirements.txt
fastapi
uvicorn
python-dotenv

langchain==0.1.20
langchain-community==0.0.38
faiss-cpu
openai